perf(bfloat16): marshal exponent functions through float; add IEEE special-case tests (#1017)
Addresses PR #1015 review feedback on the new bfloat16 exponent-manipulation
functions (ldexp/frexp/scalbn/logb/ilogb, issue #941).

- Maintainer (Ravenwater): marshal through float instead of double. A bfloat16
  is the high 16 bits of an IEEE float32 and shares its 8-bit exponent field, so
  bfloat16 -> float is exact and covers bfloat16's entire exponent range. float
  is therefore the natural, cheaper marshalling type for exponent manipulation
  (no widening to double); results are identical, only the conversion is cheaper.
- CodeRabbit: extend the test corpus with IEEE special cases. Added a section
  exercising +0/-0, +/-inf, and NaN: frexp/ldexp/scalbn preserve the value class,
  and ilogb returns the std sentinels (FP_ILOGB0, INT_MAX, FP_ILOGBNAN).

Verified gcc+clang.

Relates to #941

+++ include/sw/universal/number/bfloat16/math/functions/exponent.hpp
@@ -29,34 +29,37 @@ inline bfloat16 expm1(bfloat16 x) {
 }
 
 // ---- exponent manipulation (issue #941) ----
-// bfloat16 delegates to the host double routines (matching the exp/log family
-// above); a bfloat16 value is exactly representable in double, so the only loss
-// is the final round of the result back to bfloat16's 8-bit significand.
+// These delegate to the host FLOAT routines. A bfloat16 is the high 16 bits of
+// an IEEE-754 float32 and shares its 8-bit exponent field, so bfloat16 -> float
+// is exact and covers bfloat16's entire exponent range. float is therefore the
+// natural -- and cheaper -- marshalling type for exponent manipulation (no
+// widening to double, per PR #1015 review); the only loss is the final round of
+// the result back to bfloat16's 8-bit significand.
 
 // Decompose x into a normalized fraction in [0.5, 1) and an integer power of two:
 // x == frexp(x, &e) * 2^e. Writes the exponent to *exp.
 inline bfloat16 frexp(bfloat16 x, int* exp) {
-	return bfloat16(std::frexp(double(x), exp));
+	return bfloat16(std::frexp(float(x), exp));
 }
 
 // Multiply x by 2^n.
 inline bfloat16 ldexp(bfloat16 x, int n) {
-	return bfloat16(std::ldexp(double(x), n));
+	return bfloat16(std::ldexp(float(x), n));
 }
 
 // Multiply x by FLT_RADIX^n; FLT_RADIX == 2 here, so identical to ldexp.
 inline bfloat16 scalbn(bfloat16 x, int n) {
-	return bfloat16(std::scalbn(double(x), n));
+	return bfloat16(std::scalbn(float(x), n));
 }
 
 // Unbiased radix-2 exponent of x, as a floating-point value.
 inline bfloat16 logb(bfloat16 x) {
-	return bfloat16(std::logb(double(x)));
+	return bfloat16(std::logb(float(x)));
 }
 
 // Unbiased radix-2 exponent of x, as an int.
 inline int ilogb(bfloat16 x) {
-	return std::ilogb(double(x));
+	return std::ilogb(float(x));
 }
 
 }} // namespace sw::universal
